Create sourceset documentable filter for javadoc
