+++ plugins/base/src/main/kotlin/transformers/documentables/DocumentableVisibilityFilterTransformer.kt
@@ -5,6 +5,7 @@ import org.jetbrains.dokka.model.*
 import org.jetbrains.dokka.plugability.DokkaContext
 import org.jetbrains.dokka.transformers.documentation.PreMergeDocumentableTransformer
 import org.jetbrains.dokka.DokkaConfiguration.DokkaSourceSet
+import org.jetbrains.kotlin.utils.addToStdlib.safeAs
 
 class DocumentableVisibilityFilterTransformer(val context: DokkaContext) : PreMergeDocumentableTransformer {
 
@@ -17,7 +18,8 @@ class DocumentableVisibilityFilterTransformer(val context: DokkaContext) : PreMe
     private class DocumentableVisibilityFilter(
         val packageOptions: List<DokkaConfiguration.PackageOptions>,
         val globalOptions: DokkaSourceSet
-    ) {
+    ): AbstractDocumentableFilterTransformer {
+
         fun Visibility.isAllowedInPackage(packageName: String?) = when (this) {
             is JavaVisibility.Public,
             is JavaVisibility.Default,
@@ -27,63 +29,18 @@ class DocumentableVisibilityFilterTransformer(val context: DokkaContext) : PreMe
                     ?: globalOptions.includeNonPublic
         }
 
-        fun processModule(original: DModule) =
-            filterPackages(original.packages).let { (modified, packages) ->
-                if (!modified) original
-                else
-                    DModule(
-                        original.name,
-                        packages = packages,
-                        documentation = original.documentation,
-                        sourceSets = original.sourceSets,
-                        extra = original.extra
-                    )
-            }
-
-
-        private fun filterPackages(packages: List<DPackage>): Pair<Boolean, List<DPackage>> {
-            var packagesListChanged = false
-            val filteredPackages = packages.map {
-                var modified = false
-                val functions = filterFunctions(it.functions).let { (listModified, list) ->
-                    modified = modified || listModified
-                    list
-                }
-                val properties = filterProperties(it.properties).let { (listModified, list) ->
-                    modified = modified || listModified
-                    list
-                }
-                val classlikes = filterClasslikes(it.classlikes).let { (listModified, list) ->
-                    modified = modified || listModified
-                    list
-                }
-                when {
-                    !modified -> it
-                    else -> {
-                        packagesListChanged = true
-                        DPackage(
-                            it.dri,
-                            functions,
-                            properties,
-                            classlikes,
-                            it.typealiases,
-                            it.documentation,
-                            it.expectPresentInSet,
-                            it.sourceSets,
-                            it.extra
-                        )
-                    }
-                }
-            }
-            return Pair(packagesListChanged, filteredPackages)
-        }
-
-        private fun <T : WithVisibility> alwaysTrue(a: T, p: DokkaSourceSet) = true
-        private fun <T : WithVisibility> alwaysFalse(a: T, p: DokkaSourceSet) = false
-
         private fun WithVisibility.visibilityForPlatform(data: DokkaSourceSet): Visibility? = visibility[data]
 
-        private fun <T> T.filterPlatforms(
+        override fun <T: Documentable> T.filterPlatforms(
+            additionalCondition: (T, DokkaSourceSet) -> Boolean,
+            alternativeCondition: (T, DokkaSourceSet) -> Boolean
+        ): Set<DokkaSourceSet> =
+            when (this) {
+                is WithVisibility -> filterPlatformsWithVisibility(additionalCondition, alternativeCondition)
+                else -> emptySet()
+            }
+
+        private fun <T> T.filterPlatformsWithVisibility(
             additionalCondition: (T, DokkaSourceSet) -> Boolean = ::alwaysTrue,
             alternativeCondition: (T, DokkaSourceSet) -> Boolean = ::alwaysFalse
         ) where T : Documentable, T : WithVisibility =
@@ -93,16 +50,16 @@ class DocumentableVisibilityFilterTransformer(val context: DokkaContext) : PreMe
                         alternativeCondition(this, d)
             }.toSet()
 
-        private fun <T> List<T>.transform(
-            additionalCondition: (T, DokkaSourceSet) -> Boolean = ::alwaysTrue,
-            alternativeCondition: (T, DokkaSourceSet) -> Boolean = ::alwaysFalse,
+        override fun <T : Documentable> List<T>.transform(
+            additionalCondition: (T, DokkaSourceSet) -> Boolean,
+            alternativeCondition: (T, DokkaSourceSet) -> Boolean,
             recreate: (T, Set<DokkaSourceSet>) -> T
-        ): Pair<Boolean, List<T>> where T : Documentable, T : WithVisibility {
+        ): Pair<Boolean, List<T>> {
             var changed = false
             val values = mapNotNull { t ->
                 val filteredPlatforms = t.filterPlatforms(additionalCondition, alternativeCondition)
                 when (filteredPlatforms.size) {
-                    t.visibility.size -> t
+                    t.safeAs<WithVisibility>()?.visibility?.size -> t
                     0 -> {
                         changed = true
                         null
@@ -116,38 +73,13 @@ class DocumentableVisibilityFilterTransformer(val context: DokkaContext) : PreMe
             return Pair(changed, values)
         }
 
-        private fun filterFunctions(
-            functions: List<DFunction>,
-            additionalCondition: (DFunction, DokkaSourceSet) -> Boolean = ::alwaysTrue
-        ) =
-            functions.transform(additionalCondition) { original, filteredPlatforms ->
-                with(original) {
-                    DFunction(
-                        dri,
-                        name,
-                        isConstructor,
-                        parameters,
-                        documentation.filtered(filteredPlatforms),
-                        expectPresentInSet.filtered(filteredPlatforms),
-                        sources.filtered(filteredPlatforms),
-                        visibility.filtered(filteredPlatforms),
-                        type,
-                        generics.mapNotNull { it.filter(filteredPlatforms) },
-                        receiver,
-                        modifier,
-                        filteredPlatforms,
-                        extra
-                    )
-                }
-            }
-
         private fun hasVisibleAccessorsForPlatform(property: DProperty, data: DokkaSourceSet) =
             property.getter?.visibilityForPlatform(data)?.isAllowedInPackage(property.dri.packageName) == true ||
                     property.setter?.visibilityForPlatform(data)?.isAllowedInPackage(property.dri.packageName) == true
 
-        private fun filterProperties(
+        override fun filterProperties(
             properties: List<DProperty>,
-            additionalCondition: (DProperty, DokkaSourceSet) -> Boolean = ::alwaysTrue
+            additionalCondition: (DProperty, DokkaSourceSet) -> Boolean
         ): Pair<Boolean, List<DProperty>> =
             properties.transform(additionalCondition, ::hasVisibleAccessorsForPlatform) { original, filteredPlatforms ->
                 with(original) {
@@ -169,159 +101,5 @@ class DocumentableVisibilityFilterTransformer(val context: DokkaContext) : PreMe
                     )
                 }
             }
-
-        private fun filterEnumEntries(entries: List<DEnumEntry>, filteredPlatforms: Set<DokkaSourceSet>) =
-            entries.mapNotNull { entry ->
-                if (filteredPlatforms.containsAll(entry.sourceSets)) entry
-                else {
-                    val intersection = filteredPlatforms.intersect(entry.sourceSets)
-                    if (intersection.isEmpty()) null
-                    else DEnumEntry(
-                        entry.dri,
-                        entry.name,
-                        entry.documentation.filtered(intersection),
-                        entry.expectPresentInSet.filtered(filteredPlatforms),
-                        filterFunctions(entry.functions) { _, data -> data in intersection }.second,
-                        filterProperties(entry.properties) { _, data -> data in intersection }.second,
-                        filterClasslikes(entry.classlikes) { _, data -> data in intersection }.second,
-                        intersection,
-                        entry.extra
-                    )
-                }
-            }
-
-        private fun filterClasslikes(
-            classlikeList: List<DClasslike>,
-            additionalCondition: (DClasslike, DokkaSourceSet) -> Boolean = ::alwaysTrue
-        ): Pair<Boolean, List<DClasslike>> {
-            var classlikesListChanged = false
-            val filteredClasslikes: List<DClasslike> = classlikeList.mapNotNull {
-                with(it) {
-                    val filteredPlatforms = filterPlatforms(additionalCondition)
-                    if (filteredPlatforms.isEmpty()) {
-                        classlikesListChanged = true
-                        null
-                    } else {
-                        var modified = sourceSets.size != filteredPlatforms.size
-                        val functions =
-                            filterFunctions(functions) { _, data -> data in filteredPlatforms }.let { (listModified, list) ->
-                                modified = modified || listModified
-                                list
-                            }
-                        val properties =
-                            filterProperties(properties) { _, data -> data in filteredPlatforms }.let { (listModified, list) ->
-                                modified = modified || listModified
-                                list
-                            }
-                        val classlikes =
-                            filterClasslikes(classlikes) { _, data -> data in filteredPlatforms }.let { (listModified, list) ->
-                                modified = modified || listModified
-                                list
-                            }
-                        val companion =
-                            if (this is WithCompanion) filterClasslikes(listOfNotNull(companion)) { _, data -> data in filteredPlatforms }.let { (listModified, list) ->
-                                modified = modified || listModified
-                                list.firstOrNull() as DObject?
-                            } else null
-                        val constructors = if (this is WithConstructors)
-                            filterFunctions(constructors) { _, data -> data in filteredPlatforms }.let { (listModified, list) ->
-                                modified = modified || listModified
-                                list
-                            } else emptyList()
-                        val generics =
-                            if (this is WithGenerics) generics.mapNotNull { param -> param.filter(filteredPlatforms) } else emptyList()
-                        val enumEntries =
-                            if (this is DEnum) filterEnumEntries(entries, filteredPlatforms) else emptyList()
-                        classlikesListChanged = classlikesListChanged || modified
-                        when {
-                            !modified -> this
-                            this is DClass -> DClass(
-                                dri,
-                                name,
-                                constructors,
-                                functions,
-                                properties,
-                                classlikes,
-                                sources.filtered(filteredPlatforms),
-                                visibility.filtered(filteredPlatforms),
-                                companion,
-                                generics,
-                                supertypes.filtered(filteredPlatforms),
-                                documentation.filtered(filteredPlatforms),
-                                expectPresentInSet.filtered(filteredPlatforms),
-                                modifier,
-                                filteredPlatforms,
-                                extra
-                            )
-                            this is DAnnotation -> DAnnotation(
-                                name,
-                                dri,
-                                documentation.filtered(filteredPlatforms),
-                                expectPresentInSet.filtered(filteredPlatforms),
-                                sources.filtered(filteredPlatforms),
-                                functions,
-                                properties,
-                                classlikes,
-                                visibility.filtered(filteredPlatforms),
-                                companion,
-                                constructors,
-                                generics,
-                                filteredPlatforms,
-                                extra
-                            )
-                            this is DEnum -> DEnum(
-                                dri,
-                                name,
-                                enumEntries,
-                                documentation.filtered(filteredPlatforms),
-                                expectPresentInSet.filtered(filteredPlatforms),
-                                sources.filtered(filteredPlatforms),
-                                functions,
-                                properties,
-                                classlikes,
-                                visibility.filtered(filteredPlatforms),
-                                companion,
-                                constructors,
-                                supertypes.filtered(filteredPlatforms),
-                                filteredPlatforms,
-                                extra
-                            )
-                            this is DInterface -> DInterface(
-                                dri,
-                                name,
-                                documentation.filtered(filteredPlatforms),
-                                expectPresentInSet.filtered(filteredPlatforms),
-                                sources.filtered(filteredPlatforms),
-                                functions,
-                                properties,
-                                classlikes,
-                                visibility.filtered(filteredPlatforms),
-                                companion,
-                                generics,
-                                supertypes.filtered(filteredPlatforms),
-                                filteredPlatforms,
-                                extra
-                            )
-                            this is DObject -> DObject(
-                                name,
-                                dri,
-                                documentation.filtered(filteredPlatforms),
-                                expectPresentInSet.filtered(filteredPlatforms),
-                                sources.filtered(filteredPlatforms),
-                                functions,
-                                properties,
-                                classlikes,
-                                visibility,
-                                supertypes.filtered(filteredPlatforms),
-                                filteredPlatforms,
-                                extra
-                            )
-                            else -> null
-                        }
-                    }
-                }
-            }
-            return Pair(classlikesListChanged, filteredClasslikes)
-        }
     }
 }
